+++ llama-index-core/llama_index/core/callbacks/__init__.py
@@ -1,9 +1,6 @@
-from llama_index.core.events.base_event import CBEvent
-from llama_index.core.events.base_event_type import CBEventType
-
 from .base import CallbackManager
 from .llama_debug import LlamaDebugHandler
-from .schema import EventPayload
+from .schema import CBEvent, CBEventType, EventPayload
 from .token_counting import TokenCountingHandler
 from .utils import trace_method
