+++ llama-index-core/llama_index/core/callbacks/schema.py
@@ -1,52 +1,54 @@
 """Base schema for callback managers."""
 
-# import uuid
+import uuid
+from abc import ABC, abstractmethod
 from dataclasses import dataclass
-
-# from datetime import datetime
+from datetime import datetime
 from enum import Enum
+from typing import Any
 
-from llama_index.core.events.base_event_type import CBEventType
+from deprecated import deprecated
 
-# from typing import Any, Dict, Optional
-
+from llama_index.core.bridge.pydantic import BaseModel
 
 # timestamp for callback events
-# TIMESTAMP_FORMAT = "%m/%d/%Y, %H:%M:%S.%f"
+TIMESTAMP_FORMAT = "%m/%d/%Y, %H:%M:%S.%f"
 
 # base trace_id for the tracemap in callback_manager
 BASE_TRACE_EVENT = "root"
 
 
-# class CBEventType(str, Enum):
-#     """Callback manager event types.
+@deprecated("Since each event now has its own class now")
+class CBEventType(str, Enum):
+    """Callback manager event types.
 
-#     Attributes:
-#         CHUNKING: Logs for the before and after of text splitting.
-#         NODE_PARSING: Logs for the documents and the nodes that they are parsed into.
-#         EMBEDDING: Logs for the number of texts embedded.
-#         LLM: Logs for the template and response of LLM calls.
-#         QUERY: Keeps track of the start and end of each query.
-#         RETRIEVE: Logs for the nodes retrieved for a query.
-#         SYNTHESIZE: Logs for the result for synthesize calls.
-#         TREE: Logs for the summary and level of summaries generated.
-#         SUB_QUESTION: Logs for a generated sub question and answer.
-#     """
+    Attributes:
+        CHUNKING: Logs for the before and after of text splitting.
+        NODE_PARSING: Logs for the documents and the nodes that they are parsed into.
+        EMBEDDING: Logs for the number of texts embedded.
+        LLM: Logs for the template and response of LLM calls.
+        QUERY: Keeps track of the start and end of each query.
+        RETRIEVE: Logs for the nodes retrieved for a query.
+        SYNTHESIZE: Logs for the result for synthesize calls.
+        TREE: Logs for the summary and level of summaries generated.
+        SUB_QUESTION: Logs for a generated sub question and answer.
+    """
 
-#     CHUNKING = "chunking"
-#     NODE_PARSING = "node_parsing"
-#     EMBEDDING = "embedding"
-#     LLM = "llm"
-#     QUERY = "query"
-#     RETRIEVE = "retrieve"
-#     SYNTHESIZE = "synthesize"
-#     TREE = "tree"
-#     SUB_QUESTION = "sub_question"
-#     TEMPLATING = "templating"
-#     FUNCTION_CALL = "function_call"
-#     RERANKING = "reranking"
-#     EXCEPTION = "exception"
-#     AGENT_STEP = "agent_step"
+    CHUNKING = "chunking"
+    NODE_PARSING = "node_parsing"
+    EMBEDDING = "embedding"
+    LLM = "llm"
+    QUERY = "query"
+    RETRIEVE = "retrieve"
+    SYNTHESIZE = "synthesize"
+    TREE = "tree"
+    SUB_QUESTION = "sub_question"
+    TEMPLATING = "templating"
+    FUNCTION_CALL = "function_call"
+    RERANKING = "reranking"
+    EXCEPTION = "exception"
+    AGENT_STEP = "agent_step"
+    CUSTOM_EVENT = "custom_event"
 
 
 class EventPayload(str, Enum):
@@ -78,21 +80,40 @@ class EventPayload(str, Enum):
 LEAF_EVENTS = (CBEventType.CHUNKING, CBEventType.LLM, CBEventType.EMBEDDING)
 
 
-# @dataclass
-# class CBEvent:
-#     """Generic class to store event information."""
+class CBEvent(BaseModel, ABC):
+    """Base class to store event information."""
 
-#     event_type: CBEventType
-#     payload: Optional[Dict[str, Any]] = None
-#     time: str = ""
-#     id_: str = ""
+    _event_type: CBEventType
+    _time: str = ""
+    _id: str = ""
 
-#     def __post_init__(self) -> None:
-#         """Init time and id if needed."""
-#         if not self.time:
-#             self.time = datetime.now().strftime(TIMESTAMP_FORMAT)
-#         if not self.id_:
-#             self.id = str(uuid.uuid4())
+    def __init__(self, event_type: CBEventType) -> None:
+        """Init time and id if needed."""
+        self._event_type = event_type
+        self._time = datetime.now().strftime(TIMESTAMP_FORMAT)
+        self._id = str(uuid.uuid4())
+
+    @property
+    def id_(self) -> str:
+        """Return the UUID id for the event."""
+        return self._id
+
+    @property
+    def time(self) -> str:
+        """Return the timestamp for the event."""
+        return self._time
+
+    @property
+    @deprecated("You can call isinstance on the class to get the type of class/event.")
+    def event_type(self) -> CBEventType:
+        """Return the event type."""
+        return self._event_type
+
+    @property
+    @abstractmethod
+    @deprecated("You can access the payload properties directly from the class.")
+    def payload(self) -> Any:
+        """Return the payload for the event (to support legacy systems)."""
 
 
 @dataclass
