+++ llama-index-core/llama_index/core/events/agent_step_events.py
@@ -5,9 +5,8 @@ from typing import List
 
 from deprecated import deprecated
 
+from llama_index.core.callbacks.schema import CBEvent, CBEventType
 from llama_index.core.chat_engine.types import AgentChatResponse
-from llama_index.core.events.base_event import CBEvent
-from llama_index.core.events.base_event_type import CBEventType
 
 
 @dataclass
