--- llama-index-core/llama_index/core/events/base_event_type.py
@@ -1,36 +1,0 @@
-from enum import Enum
-
-from deprecated import deprecated
-
-
-@deprecated("Since each event now has its own class now")
-class CBEventType(str, Enum):
-    """Callback manager event types.
-
-    Attributes:
-        CHUNKING: Logs for the before and after of text splitting.
-        NODE_PARSING: Logs for the documents and the nodes that they are parsed into.
-        EMBEDDING: Logs for the number of texts embedded.
-        LLM: Logs for the template and response of LLM calls.
-        QUERY: Keeps track of the start and end of each query.
-        RETRIEVE: Logs for the nodes retrieved for a query.
-        SYNTHESIZE: Logs for the result for synthesize calls.
-        TREE: Logs for the summary and level of summaries generated.
-        SUB_QUESTION: Logs for a generated sub question and answer.
-    """
-
-    CHUNKING = "chunking"
-    NODE_PARSING = "node_parsing"
-    EMBEDDING = "embedding"
-    LLM = "llm"
-    QUERY = "query"
-    RETRIEVE = "retrieve"
-    SYNTHESIZE = "synthesize"
-    TREE = "tree"
-    SUB_QUESTION = "sub_question"
-    TEMPLATING = "templating"
-    FUNCTION_CALL = "function_call"
-    RERANKING = "reranking"
-    EXCEPTION = "exception"
-    AGENT_STEP = "agent_step"
-    CUSTOM_EVENT = "custom_event"
