+++ llama-index-core/llama_index/core/events/sub_question_events.py
@@ -4,8 +4,7 @@ from dataclasses import dataclass
 
 from deprecated import deprecated
 
-from llama_index.core.events.base_event import CBEvent
-from llama_index.core.events.base_event_type import CBEventType
+from llama_index.core.callbacks.schema import CBEvent, CBEventType
 from llama_index.core.query_engine.sub_question_query_engine import (
     SubQuestionAnswerPair,
 )
